// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  SALES: 'SALES',
  TEAM_LEAD: 'TEAM_LEAD',
  MANAGER: 'MANAGER',
  OWNER_FINANCE: 'OWNER_FINANCE',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const IncentiveStatus = {
  PENDING_REVIEW: 'PENDING_REVIEW',
  ON_HOLD: 'ON_HOLD',
  CLAIMABLE: 'CLAIMABLE',
  CLAIM_REQUESTED: 'CLAIM_REQUESTED',
  PAID: 'PAID'
} as const

export type IncentiveStatus = (typeof IncentiveStatus)[keyof typeof IncentiveStatus]


export const IncentiveLevel = {
  L1: 'L1',
  L2: 'L2',
  L3: 'L3',
  L4: 'L4'
} as const

export type IncentiveLevel = (typeof IncentiveLevel)[keyof typeof IncentiveLevel]


export const IncentiveRuleType = {
  PERCENTAGE: 'PERCENTAGE'
} as const

export type IncentiveRuleType = (typeof IncentiveRuleType)[keyof typeof IncentiveRuleType]


export const PropertyType = {
  APARTMENT: 'APARTMENT',
  VILLA: 'VILLA',
  PLOT: 'PLOT',
  COMMERCIAL: 'COMMERCIAL'
} as const

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType]


export const BrokerChannel = {
  DIRECT: 'DIRECT',
  PARTNER: 'PARTNER',
  BROKER: 'BROKER'
} as const

export type BrokerChannel = (typeof BrokerChannel)[keyof typeof BrokerChannel]